finished migrating all fixtures to new folder
