+++ tests/fixtures/state/stores/transactions.py
@@ -1,0 +1,8 @@
+import pytest
+
+from quixstreams.state.base.transaction import PartitionTransactionCache
+
+
+@pytest.fixture()
+def cache() -> PartitionTransactionCache:
+    return PartitionTransactionCache()
