+++ tests/test_quixstreams/test_checkpointing.py
@@ -1,18 +1,15 @@
 import contextlib
-from typing import Optional
 from unittest.mock import MagicMock, patch
 
 import pytest
 from confluent_kafka import KafkaError, KafkaException, TopicPartition
 
-from quixstreams.checkpointing import Checkpoint, InvalidStoredOffset
+from quixstreams.checkpointing import InvalidStoredOffset
 from quixstreams.checkpointing.exceptions import (
     CheckpointConsumerCommitError,
     CheckpointProducerTimeout,
 )
 from quixstreams.kafka import Consumer
-from quixstreams.processing import PausingManager
-from quixstreams.rowproducer import RowProducer
 from quixstreams.sinks import BatchingSink, SinkBackpressureError, SinkManager
 from quixstreams.sinks.base import SinkBatch
 from quixstreams.state import StateStoreManager
@@ -20,44 +17,6 @@ from quixstreams.state.base import PartitionTransaction
 from quixstreams.state.exceptions import StoreNotRegisteredError, StoreTransactionFailed
 from quixstreams.state.manager import SUPPORTED_STORES
 from tests.utilities.utils import DummySink
-
-
-@pytest.fixture()
-def checkpoint_factory(state_manager, consumer, row_producer_factory):
-    def factory(
-        commit_interval: float = 1,
-        commit_every: int = 0,
-        consumer_: Optional[Consumer] = None,
-        producer_: Optional[RowProducer] = None,
-        state_manager_: Optional[StateStoreManager] = None,
-        sink_manager_: Optional[SinkManager] = None,
-        pausing_manager_: Optional[PausingManager] = None,
-        exactly_once: bool = False,
-    ):
-        consumer_ = consumer_ or consumer
-        sink_manager_ = sink_manager_ or SinkManager()
-        pausing_manager_ = pausing_manager_ or PausingManager(consumer=consumer)
-        producer_ = producer_ or row_producer_factory(transactional=exactly_once)
-        state_manager_ = state_manager_ or state_manager
-        return Checkpoint(
-            commit_interval=commit_interval,
-            commit_every=commit_every,
-            producer=producer_,
-            consumer=consumer_,
-            state_manager=state_manager_,
-            sink_manager=sink_manager_,
-            pausing_manager=pausing_manager_,
-            exactly_once=exactly_once,
-        )
-
-    return factory
-
-
-@pytest.fixture()
-def rowproducer_mock(request):
-    p = MagicMock(spec_set=RowProducer)
-    p.flush.return_value = getattr(request, "param", 0)
-    return p
 
 
 class BackpressuredSink(BatchingSink):
@@ -136,13 +95,15 @@ class TestCheckpoint:
         consumer,
         state_manager_factory,
         topic_factory,
-        rowproducer_mock,
+        row_producer_mock,
     ):
         topic_name, _ = topic_factory()
 
-        state_manager = state_manager_factory(producer=rowproducer_mock)
+        state_manager = state_manager_factory(producer=row_producer_mock)
         checkpoint = checkpoint_factory(
-            consumer_=consumer, state_manager_=state_manager, producer_=rowproducer_mock
+            consumer_=consumer,
+            state_manager_=state_manager,
+            producer_=row_producer_mock,
         )
         processed_offset = 999
         key, value, prefix = "key", "value", b"__key__"
@@ -163,7 +124,7 @@ class TestCheckpoint:
         assert tp.offset == processed_offset + 1
 
         # Check the producer is flushed
-        assert rowproducer_mock.flush.call_count == 1
+        assert row_producer_mock.flush.call_count == 1
 
         # Check the state is flushed
         assert tx.completed
@@ -263,7 +224,7 @@ class TestCheckpoint:
     def test_close_no_offsets(
         self,
         checkpoint_factory,
-        rowproducer_mock,
+        row_producer_mock,
         exactly_once,
     ):
         consumer_mock = MagicMock(spec_set=Consumer)
@@ -271,7 +232,7 @@ class TestCheckpoint:
         checkpoint = checkpoint_factory(
             consumer_=consumer_mock,
             state_manager_=state_manager,
-            producer_=rowproducer_mock,
+            producer_=row_producer_mock,
             exactly_once=exactly_once,
         )
         # Commit the checkpoint without processing any messages
@@ -279,9 +240,9 @@ class TestCheckpoint:
 
         if exactly_once:
             # transaction should also be aborted
-            assert rowproducer_mock.abort_transaction.call_count
+            assert row_producer_mock.abort_transaction.call_count
         else:
-            assert not rowproducer_mock.abort_transaction.call_count
+            assert not row_producer_mock.abort_transaction.call_count
 
     @pytest.mark.parametrize("exactly_once", [False, True])
     def test_commit_has_failed_transactions_fails(
@@ -289,15 +250,15 @@ class TestCheckpoint:
         checkpoint_factory,
         state_manager_factory,
         topic_factory,
-        rowproducer_mock,
+        row_producer_mock,
         exactly_once,
     ):
         consumer_mock = MagicMock(spec_set=Consumer)
-        state_manager = state_manager_factory(producer=rowproducer_mock)
+        state_manager = state_manager_factory(producer=row_producer_mock)
         checkpoint = checkpoint_factory(
             consumer_=consumer_mock,
             state_manager_=state_manager,
-            producer_=rowproducer_mock,
+            producer_=row_producer_mock,
             exactly_once=exactly_once,
         )
         processed_offset = 999
@@ -328,10 +289,10 @@ class TestCheckpoint:
             checkpoint.commit()
 
         # The producer should not flush
-        assert not rowproducer_mock.flush.call_count
+        assert not row_producer_mock.flush.call_count
 
         # Check nothing is committed
-        assert not rowproducer_mock.commit_transaction.call_count
+        assert not row_producer_mock.commit_transaction.call_count
         assert not consumer_mock.commit.call_count
 
     @pytest.mark.parametrize("exactly_once", [False, True])
@@ -340,15 +301,15 @@ class TestCheckpoint:
         checkpoint_factory,
         state_manager_factory,
         topic_factory,
-        rowproducer_mock,
+        row_producer_mock,
         exactly_once,
     ):
         consumer_mock = MagicMock(spec_set=Consumer)
-        state_manager = state_manager_factory(producer=rowproducer_mock)
+        state_manager = state_manager_factory(producer=row_producer_mock)
         checkpoint = checkpoint_factory(
             consumer_=consumer_mock,
             state_manager_=state_manager,
-            producer_=rowproducer_mock,
+            producer_=row_producer_mock,
             exactly_once=exactly_once,
         )
         processed_offset = 999
@@ -362,27 +323,31 @@ class TestCheckpoint:
         tx.set(key=key, value=value, prefix=prefix)
         checkpoint.store_offset("topic", 0, processed_offset)
 
-        rowproducer_mock.flush.side_effect = ValueError("Flush failure")
+        row_producer_mock.flush.side_effect = ValueError("Flush failure")
         # Checkpoint commit should fail if producer failed to flush
         with pytest.raises(ValueError):
             checkpoint.commit()
 
         # Nothing should commit
-        assert not rowproducer_mock.commit_transaction.call_count
+        assert not row_producer_mock.commit_transaction.call_count
         assert not consumer_mock.commit.call_count
         # The transaction should remain prepared, but not completed
         assert tx.prepared
         assert not tx.completed
 
     def test_commit_consumer_commit_fails(
-        self, checkpoint_factory, state_manager_factory, topic_factory, rowproducer_mock
+        self,
+        checkpoint_factory,
+        state_manager_factory,
+        topic_factory,
+        row_producer_mock,
     ):
         consumer_mock = MagicMock(spec_set=Consumer)
-        state_manager = state_manager_factory(producer=rowproducer_mock)
+        state_manager = state_manager_factory(producer=row_producer_mock)
         checkpoint = checkpoint_factory(
             consumer_=consumer_mock,
             state_manager_=state_manager,
-            producer_=rowproducer_mock,
+            producer_=row_producer_mock,
         )
         processed_offset = 999
         key, value, prefix = "key", "value", b"__key__"
@@ -401,7 +366,7 @@ class TestCheckpoint:
             checkpoint.commit()
 
         # Producer should flush
-        assert rowproducer_mock.flush.call_count
+        assert row_producer_mock.flush.call_count
         # The transaction should remain prepared, but not completed
         assert tx.prepared
         assert not tx.completed
@@ -423,13 +388,15 @@ class TestCheckpoint:
         tx2 = checkpoint.get_store_transaction("topic", 0, "default")
         assert tx2 is tx
 
-    @pytest.mark.parametrize("rowproducer_mock", [1], indirect=True)
+    @pytest.mark.parametrize("row_producer_mock", [1], indirect=True)
     def test_incomplete_flush(
-        self, checkpoint_factory, consumer, state_manager_factory, rowproducer_mock
+        self, checkpoint_factory, consumer, state_manager_factory, row_producer_mock
     ):
-        state_manager = state_manager_factory(producer=rowproducer_mock)
+        state_manager = state_manager_factory(producer=row_producer_mock)
         checkpoint = checkpoint_factory(
-            consumer_=consumer, state_manager_=state_manager, producer_=rowproducer_mock
+            consumer_=consumer,
+            state_manager_=state_manager,
+            producer_=row_producer_mock,
         )
         checkpoint.store_offset("topic", 0, 0)
 
@@ -442,16 +409,16 @@ class TestCheckpoint:
         )
 
     def test_failed_commit(
-        self, checkpoint_factory, state_manager_factory, rowproducer_mock
+        self, checkpoint_factory, state_manager_factory, row_producer_mock
     ):
         consumer_mock = MagicMock(spec_set=Consumer)
         consumer_mock.commit.side_effect = KafkaException(KafkaError(1, "test error"))
 
-        state_manager = state_manager_factory(producer=rowproducer_mock)
+        state_manager = state_manager_factory(producer=row_producer_mock)
         checkpoint = checkpoint_factory(
             consumer_=consumer_mock,
             state_manager_=state_manager,
-            producer_=rowproducer_mock,
+            producer_=row_producer_mock,
         )
         checkpoint.store_offset("topic", 0, 0)
 
@@ -464,7 +431,7 @@ class TestCheckpoint:
         )
 
     def test_failed_commit_partition(
-        self, checkpoint_factory, state_manager_factory, rowproducer_mock
+        self, checkpoint_factory, state_manager_factory, row_producer_mock
     ):
         consumer_mock = MagicMock(spec_set=Consumer)
 
@@ -472,11 +439,11 @@ class TestCheckpoint:
         topic_partition.error = KafkaError(1, "test error")
 
         consumer_mock.commit.return_value = [topic_partition]
-        state_manager = state_manager_factory(producer=rowproducer_mock)
+        state_manager = state_manager_factory(producer=row_producer_mock)
         checkpoint = checkpoint_factory(
             consumer_=consumer_mock,
             state_manager_=state_manager,
-            producer_=rowproducer_mock,
+            producer_=row_producer_mock,
         )
         checkpoint.store_offset("topic", 0, 0)
 
@@ -495,7 +462,7 @@ class TestCheckpoint:
         state_manager,
         checkpoint_factory,
         state_manager_factory,
-        rowproducer_mock,
+        row_producer_mock,
     ):
         topic_name, _ = topic_factory()
         sink_manager = SinkManager()
@@ -541,7 +508,7 @@ class TestCheckpoint:
         state_manager,
         checkpoint_factory,
         state_manager_factory,
-        rowproducer_mock,
+        row_producer_mock,
     ):
         topic_name, _ = topic_factory()
         sink_manager = SinkManager()
@@ -587,7 +554,7 @@ class TestCheckpoint:
         state_manager,
         checkpoint_factory,
         state_manager_factory,
-        rowproducer_mock,
+        row_producer_mock,
     ):
         topic_name, _ = topic_factory()
         sink_manager = SinkManager()
