+++ tests/test_quixstreams/test_dataframe/test_windows/test_tumbling.py
@@ -3,18 +3,6 @@ import pytest
 from quixstreams.dataframe.windows import TumblingWindowDefinition
 
 
-@pytest.fixture()
-def tumbling_window_definition_factory(state_manager, dataframe_factory):
-    def factory(duration_ms: int, grace_ms: int = 0) -> TumblingWindowDefinition:
-        sdf = dataframe_factory(state_manager=state_manager)
-        window_def = TumblingWindowDefinition(
-            duration_ms=duration_ms, grace_ms=grace_ms, dataframe=sdf
-        )
-        return window_def
-
-    return factory
-
-
 class TestTumblingWindow:
     @pytest.mark.parametrize(
         "duration, grace, provided_name, func_name, expected_name",
