+++ tests/test_quixstreams/test_state/test_rocksdb/test_windowed/test_state.py
@@ -1,54 +1,35 @@
-from contextlib import contextmanager
-
 import pytest
 
 
-@pytest.fixture
-def store(windowed_rocksdb_store_factory):
-    store = windowed_rocksdb_store_factory()
-    store.assign_partition(0)
-    return store
-
-
-@pytest.fixture
-def transaction_state(store):
-    @contextmanager
-    def _transaction_state():
-        with store.start_partition_transaction(0) as tx:
-            yield tx.as_state(prefix=b"__key__")
-
-    return _transaction_state
-
-
-def test_update_window(transaction_state):
-    with transaction_state() as state:
+def test_update_window(windowed_rocksdb_transaction_state):
+    with windowed_rocksdb_transaction_state() as state:
         state.update_window(start_ms=0, end_ms=10, value=1, timestamp_ms=2)
         assert state.get_window(start_ms=0, end_ms=10) == 1
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         assert state.get_window(start_ms=0, end_ms=10) == 1
 
 
-def test_update_window_with_window_timestamp(transaction_state):
-    with transaction_state() as state:
+def test_update_window_with_window_timestamp(windowed_rocksdb_transaction_state):
+    with windowed_rocksdb_transaction_state() as state:
         state.update_window(
             start_ms=0, end_ms=10, value=1, timestamp_ms=2, window_timestamp_ms=3
         )
         assert state.get_window(start_ms=0, end_ms=10) == [3, 1]
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         assert state.get_window(start_ms=0, end_ms=10) == [3, 1]
 
 
 @pytest.mark.parametrize("delete", [True, False])
-def test_expire_windows(transaction_state, delete):
+def test_expire_windows(windowed_rocksdb_transaction_state, delete):
     duration_ms = 10
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         state.update_window(start_ms=0, end_ms=10, value=1, timestamp_ms=2)
         state.update_window(start_ms=10, end_ms=20, value=2, timestamp_ms=10)
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         state.update_window(start_ms=20, end_ms=30, value=3, timestamp_ms=20)
         max_start_time = state.get_latest_timestamp() - duration_ms
         expired = state.expire_windows(max_start_time=max_start_time, delete=delete)
@@ -62,19 +43,19 @@ def test_expire_windows(transaction_state, delete):
         ((10, 20), 2),
     ]
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         assert state.get_window(start_ms=0, end_ms=10) == None if delete else 1
         assert state.get_window(start_ms=10, end_ms=20) == None if delete else 2
         assert state.get_window(start_ms=20, end_ms=30) == 3
 
 
-def test_same_keys_in_db_and_update_cache(transaction_state):
+def test_same_keys_in_db_and_update_cache(windowed_rocksdb_transaction_state):
     duration_ms = 10
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         state.update_window(start_ms=0, end_ms=10, value=1, timestamp_ms=2)
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         # The same window already exists in the db
         state.update_window(start_ms=0, end_ms=10, value=3, timestamp_ms=8)
 
@@ -252,13 +233,13 @@ def test_get_windows(
     deleted_windows,
     get_windows_args,
     expected_windows,
-    transaction_state,
+    windowed_rocksdb_transaction_state,
 ):
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         for window in db_windows:
             state.update_window(**window)
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         for window in cached_windows:
             state.update_window(**window)
         for window in deleted_windows:
@@ -268,13 +249,13 @@ def test_get_windows(
         assert list(windows) == expected_windows
 
 
-def test_delete_windows(transaction_state):
-    with transaction_state() as state:
+def test_delete_windows(windowed_rocksdb_transaction_state):
+    with windowed_rocksdb_transaction_state() as state:
         state.update_window(start_ms=1, end_ms=2, value=1, timestamp_ms=1)
         state.update_window(start_ms=2, end_ms=3, value=2, timestamp_ms=2)
         state.update_window(start_ms=3, end_ms=4, value=3, timestamp_ms=3)
 
-    with transaction_state() as state:
+    with windowed_rocksdb_transaction_state() as state:
         assert state.get_window(start_ms=1, end_ms=2)
         assert state.get_window(start_ms=2, end_ms=3)
         assert state.get_window(start_ms=3, end_ms=4)
